Implement LGUG2Z panic removing
